Added source code documentation

+++ app/src/main/java/dev/lengua/model/EntryDatabaseModel.kt
@@ -3,8 +3,17 @@ package dev.lengua.model
 import dev.lengua.ui.Entry
 import dev.lengua.ui.IdentifiedEntry
 
+/**
+ * An entry as stored in the application database.
+ *
+ * @property id the identifier of the entry provided by the application
+ * database.
+ * @property term the text of the term saved in the application database.
+ * @property definition the text of the definition saved in the application
+ * database.
+ */
 data class EntryDatabaseModel(
-    val id: Int,
+    val id: Long,
     val term: String,
     val definition: String
 ) {
