+++ app/src/main/java/dev/lengua/ui/IdentifiedEntry.kt
@@ -1,10 +1,20 @@
 package dev.lengua.ui
 
+/**
+ * An entry with an identifier.
+ *
+ * The identifier is a long integer that uniquely identifies an entry.
+ *
+ * @property id the unique identifier of this entry.
+ * @property term the term of this entry.
+ * @property definition the definition of this entry.
+ */
 data class IdentifiedEntry(
-    private val id: Int,
+    private val id: Long,
     val term: String,
-    val definition: String)
-{
+    val definition: String
+) {
+    // The entry to identify
     private val entry = Entry(term, definition)
 
     fun id() = id
